Split BrowserWebSocketHandler out into its own file
