+++ lib/streamlit/web/server/server.py
@@ -13,17 +13,13 @@
 # limitations under the License.
 
 import asyncio
-import base64
-import binascii
+import errno
 import logging
 import os
 import socket
 import sys
-import errno
-import json
 import time
 import traceback
-import click
 from enum import Enum
 from typing import (
     Any,
@@ -31,52 +27,42 @@ from typing import (
     Optional,
     Tuple,
     Callable,
-    Awaitable,
     List,
-    Union,
 )
 
+import click
 import tornado.concurrent
 import tornado.ioloop
 import tornado.locks
 import tornado.netutil
 import tornado.web
 import tornado.websocket
-from tornado.platform.asyncio import AsyncIOLoop
-from tornado.websocket import WebSocketHandler
 from tornado.httpserver import HTTPServer
-from typing_extensions import Protocol
+from tornado.platform.asyncio import AsyncIOLoop
 
 from streamlit import config
 from streamlit import file_util
 from streamlit import source_util
 from streamlit import util
+from streamlit.app_session import AppSession
 from streamlit.caching import get_memo_stats_provider, get_singleton_stats_provider
+from streamlit.components.v1.components import ComponentRegistry
 from streamlit.config_option import ConfigOption
 from streamlit.forward_msg_cache import ForwardMsgCache
 from streamlit.forward_msg_cache import create_reference_msg
 from streamlit.forward_msg_cache import populate_hash_if_needed
 from streamlit.in_memory_file_manager import in_memory_file_manager
 from streamlit.legacy_caching.caching import _mem_caches
-from streamlit.app_session import AppSession
-from streamlit.stats import StatsManager
-from streamlit.uploaded_file_manager import UploadedFileManager
 from streamlit.logger import get_logger
-from streamlit.components.v1.components import ComponentRegistry
-from streamlit.proto.BackMsg_pb2 import BackMsg
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
-from .stats_request_handler import StatsRequestHandler
-from .component_request_handler import ComponentRequestHandler
-from streamlit.web.server.upload_file_request_handler import (
-    UploadFileRequestHandler,
-    UPLOAD_FILE_ROUTE,
-)
-
 from streamlit.session_data import SessionData
 from streamlit.state import (
     SCRIPT_RUN_WITHOUT_ERRORS_KEY,
     SessionStateStatProvider,
 )
+from streamlit.stats import StatsManager
+from streamlit.uploaded_file_manager import UploadedFileManager
+from streamlit.watcher import LocalSourcesWatcher
 from streamlit.web.server.routes import AddSlashHandler
 from streamlit.web.server.routes import AssetsFileHandler
 from streamlit.web.server.routes import DebugHandler
@@ -84,12 +70,17 @@ from streamlit.web.server.routes import HealthHandler
 from streamlit.web.server.routes import MediaFileHandler
 from streamlit.web.server.routes import MessageCacheHandler
 from streamlit.web.server.routes import StaticFileHandler
-from streamlit.web.server.server_util import is_cacheable_msg
-from streamlit.web.server.server_util import is_url_from_allowed_origins
-from streamlit.web.server.server_util import make_url_path_regex
-from streamlit.web.server.server_util import serialize_forward_msg
 from streamlit.web.server.server_util import get_max_message_size_bytes
-from streamlit.watcher import LocalSourcesWatcher
+from streamlit.web.server.server_util import is_cacheable_msg
+from streamlit.web.server.server_util import make_url_path_regex
+from streamlit.web.server.upload_file_request_handler import (
+    UploadFileRequestHandler,
+    UPLOAD_FILE_ROUTE,
+)
+from .browser_websocket_handler import _BrowserWebSocketHandler
+from .component_request_handler import ComponentRequestHandler
+from .session_client import SessionClient, SessionClientDisconnectedError
+from .stats_request_handler import StatsRequestHandler
 
 LOGGER = get_logger(__name__)
 
@@ -117,21 +108,6 @@ UNIX_SOCKET_PREFIX = "unix://"
 
 # Wait for the script run result for 60s and if no result is available give up
 SCRIPT_RUN_CHECK_TIMEOUT = 60
-
-
-class SessionClientDisconnectedError(Exception):
-    """Raised by operations on a disconnected SessionClient."""
-
-
-class SessionClient(Protocol):
-    """Interface for sending data to a session's client."""
-
-    def write_forward_msg(self, msg: ForwardMsg) -> None:
-        """Deliver a ForwardMsg to the client.
-
-        If the SessionClient has been disconnected, it should raise a
-        SessionClientDisconnectedError.
-        """
 
 
 class SessionInfo:
@@ -713,108 +689,6 @@ Please report this bug at https://github.com/streamlit/streamlit/issues.
 
         if len(self._session_info_by_id) == 0:
             self._set_state(State.NO_SESSIONS_CONNECTED)
-
-
-class _BrowserWebSocketHandler(WebSocketHandler, SessionClient):
-    """Handles a WebSocket connection from the browser"""
-
-    def initialize(self, server: Server) -> None:
-        self._server = server
-        self._session: Optional[AppSession] = None
-        # The XSRF cookie is normally set when xsrf_form_html is used, but in a pure-Javascript application
-        # that does not use any regular forms we just need to read the self.xsrf_token manually to set the
-        # cookie as a side effect.
-        # See https://www.tornadoweb.org/en/stable/guide/security.html#cross-site-request-forgery-protection
-        # for more details.
-        if config.get_option("server.enableXsrfProtection"):
-            _ = self.xsrf_token
-
-    def check_origin(self, origin: str) -> bool:
-        """Set up CORS."""
-        return super().check_origin(origin) or is_url_from_allowed_origins(origin)
-
-    def write_forward_msg(self, msg: ForwardMsg) -> None:
-        """Send a ForwardMsg to the browser."""
-        try:
-            self.write_message(serialize_forward_msg(msg), binary=True)
-        except tornado.websocket.WebSocketClosedError as e:
-            raise SessionClientDisconnectedError from e
-
-    def open(self, *args, **kwargs) -> Optional[Awaitable[None]]:
-        # Extract user info from the X-Streamlit-User header
-        is_public_cloud_app = False
-
-        try:
-            header_content = self.request.headers["X-Streamlit-User"]
-            payload = base64.b64decode(header_content)
-            user_obj = json.loads(payload)
-            email = user_obj["email"]
-            is_public_cloud_app = user_obj["isPublicCloudApp"]
-        except (KeyError, binascii.Error, json.decoder.JSONDecodeError):
-            email = "test@localhost.com"
-
-        user_info: Dict[str, Optional[str]] = dict()
-        if is_public_cloud_app:
-            user_info["email"] = None
-        else:
-            user_info["email"] = email
-
-        self._session = self._server._create_app_session(self, user_info)
-        return None
-
-    def on_close(self) -> None:
-        if not self._session:
-            return
-        self._server._close_app_session(self._session.id)
-        self._session = None
-
-    def get_compression_options(self) -> Optional[Dict[Any, Any]]:
-        """Enable WebSocket compression.
-
-        Returning an empty dict enables websocket compression. Returning
-        None disables it.
-
-        (See the docstring in the parent class.)
-        """
-        if config.get_option("server.enableWebsocketCompression"):
-            return {}
-        return None
-
-    def on_message(self, payload: Union[str, bytes]) -> None:
-        if not self._session:
-            return
-
-        msg = BackMsg()
-
-        try:
-            if isinstance(payload, str):
-                # Sanity check. (The frontend should only be sending us bytes;
-                # Protobuf.ParseFromString does not accept str input.)
-                raise RuntimeError(
-                    "WebSocket received an unexpected `str` message. "
-                    "(We expect `bytes` only.)"
-                )
-
-            msg.ParseFromString(payload)
-            LOGGER.debug("Received the following back message:\n%s", msg)
-
-            if msg.WhichOneof("type") == "close_connection":
-                # "close_connection" is a special developmentMode-only
-                # message used in e2e tests to test disabling widgets.
-                if config.get_option("global.developmentMode"):
-                    self._server.stop()
-                else:
-                    LOGGER.warning(
-                        "Client tried to close connection when "
-                        "not in development mode"
-                    )
-            else:
-                # AppSession handles all other BackMsg types.
-                self._session.handle_backmsg(msg)
-
-        except BaseException as e:
-            LOGGER.error(e)
-            self._session.handle_backmsg_exception(e)
 
 
 def _set_tornado_log_levels() -> None:
